Remove unused JavaScript bundle from dist assets
