+++ src/App.css
@@ -345,3 +345,25 @@ body {
   margin-top: 20px;
 }
 
+
+/* Location picker */
+.location-picker {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: 8px;
+  margin-top: 8px;
+}
+
+.location-label {
+  font-size: 14px;
+  color: #334155;
+}
+
+.location-select {
+  width: 100%;
+  padding: 12px;
+  border-radius: 12px;
+  border: 1px solid #e5e7eb;
+  font-size: 16px;
+}
+
